+++ pi-monitor/pi-monitor-client/browser/react/src/jsMain/kotlin/pimonitor/authentication/signup/SignUp.kt
@@ -8,17 +8,14 @@ import kotlinx.extensions.onDesktop
 import kotlinx.extensions.onMobile
 import kotlinx.extensions.text
 import kotlinx.html.InputType
-import logging.logger
 import pimonitor.PiMonitorService
 import pimonitor.authentication.signup.exports.SignUpScope
-import pimonitor.monitors.SignUpParams
 import react.Props
 import react.RBuilder
 import react.dom.p
 import react.fc
 import react.router.dom.useHistory
 import react.router.dom.withRouter
-import react.useEffectOnce
 import reakt.*
 import styled.css
 import styled.styledH2