+++ pi-monitor/pi-monitor-client/sdks/full/src/jsMain/kotlin/pimonitor/evaluation/business/BusinessesScope.kt
@@ -3,13 +3,14 @@
 
 package pimonitor.evaluation.business
 
+import pimonitor.PiMonitorService
 import pimonitor.evaluation.businesses.BusinessesService
 import viewmodel.ViewModel
 import pimonitor.evaluation.business.BusinessesIntent as Intent
 import pimonitor.evaluation.business.BusinessesState as State
 
-class BusinessesScope(service: BusinessesService) {
-    val viewModel: ViewModel<Intent, State> = BusinessViewModel(service)
+class BusinessesScope(val service: PiMonitorService) {
+    val viewModel: ViewModel<Intent, State> by lazy { BusinessViewModel(service.businesses) }
 
     val loadBusiness: () -> Unit = { viewModel.post(Intent.LoadBusinesses) }
 }