+++ pi-monitor/pi-monitor-core/src/commonMain/kotlin/pimonitor/authentication/signup/SignUpParams.kt
@@ -1,4 +1,4 @@
-package pimonitor.monitors
+package pimonitor.authentication.signup
 
 import kotlinx.serialization.Serializable