+++ pi-monitor/pi-monitor-daos/core/src/commonMain/kotlin/pimonitor/monitors/MonitorDao.kt
@@ -3,6 +3,7 @@ package pimonitor.monitors
 import bitframe.authentication.users.UserRef
 import bitframe.daos.conditions.Condition
 import later.Later
+import pimonitor.authentication.signup.SignUpParams
 
 interface MonitorDao {
     fun create(params: SignUpParams, ref: UserRef): Later<Monitor>