+++ pi-monitor/pi-monitor-services/core/src/commonMain/kotlin/pimonitor/authentication/signup/SignUpService.kt
@@ -9,7 +9,6 @@ import contacts.Email
 import later.Later
 import later.await
 import later.later
-import pimonitor.monitors.SignUpParams
 import kotlin.js.JsExport
 
 abstract class SignUpService(