+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.31
+
+## PiMonitor Client SDK Core
+
+- [[PM124]](https://github.com/picortex/bitframe/issues/124) Removed onUserLoggedIn method from SignInViewModel
+- [[PM141]](https://github.com/picortex/bitframe/issues/141) Created a Register business form 
+
 # 0.0.30
 
 ## PiMonitor Client SDK Core